+++ Strings/MurmurHash3_constexpr.inl
@@ -21,13 +21,10 @@
 //compile time murmur3_x86_32 hash (gcc)
 //https://gist.github.com/mattyclarkson/5318077
 
-constexpr inline size_t StringLengthCExpr(const char * const str) noexcept
+constexpr inline uint32_t StringLengthCExpr(const char * const str) noexcept
 {
 	return (*str == 0) ? 0 : StringLengthCExpr(str + 1) + 1;
 };
-
-
-
 
 
 constexpr inline uint32_t Murmur3Rotate_32CExpr(const uint32_t target, const uint8_t rotation) noexcept
@@ -35,35 +32,35 @@ constexpr inline uint32_t Murmur3Rotate_32CExpr(const uint32_t target, const uin
 	return (target << rotation) | (target >> (32 - rotation));
 };
 
-constexpr inline uint32_t Murmur3ShiftXor_32CExpr(const uint32_t hash, const size_t shift) noexcept
+constexpr inline uint32_t Murmur3ShiftXor_32CExpr(const uint32_t hash, const uint32_t shift) noexcept
 {
 	return hash ^ (hash >> shift);
 };
 
-constexpr inline uint32_t Murmur3Last_32CExpr(const size_t len, const uint32_t hash) noexcept
+constexpr inline uint32_t Murmur3Last_32CExpr(const uint32_t len, const uint32_t hash) noexcept
 {
 	return Murmur3ShiftXor_32CExpr(0xc2b2ae35 * Murmur3ShiftXor_32CExpr(
 		0x85ebca6b * Murmur3ShiftXor_32CExpr(
 		hash ^ len, 16), 13), 16);
 };
 
-constexpr inline uint32_t Murmur3Tail1_32CExpr(const char data, const size_t len, const uint32_t hash, const uint32_t constant) noexcept
+constexpr inline uint32_t Murmur3Tail1_32CExpr(const char data, const uint32_t len, const uint32_t hash, const uint32_t constant) noexcept
 {
 	return Murmur3Last_32CExpr(len, hash ^ (0x1b873593 * Murmur3Rotate_32CExpr(
 		0xcc9e2d51 * (constant ^ static_cast<uint8_t>(data)), 15)));
 };
 
-constexpr inline uint32_t Murmur3Tail2_32CExpr(const char data[2], const size_t len, const uint32_t hash, const uint32_t constant) noexcept
+constexpr inline uint32_t Murmur3Tail2_32CExpr(const char data[2], const uint32_t len, const uint32_t hash, const uint32_t constant) noexcept
 {
 	return Murmur3Tail1_32CExpr(data[0], len, hash, constant ^ (static_cast<uint8_t>(data[1]) << 8));
 };
 
-constexpr inline uint32_t Murmur3Tail3_32CExpr(const char data[3], const size_t len, const uint32_t hash) noexcept
+constexpr inline uint32_t Murmur3Tail3_32CExpr(const char data[3], const uint32_t len, const uint32_t hash) noexcept
 {
 	return Murmur3Tail2_32CExpr(data, len, hash, 0 ^ (static_cast<uint8_t>(data[2]) << 16));
 };
 
-constexpr inline uint32_t Murmur3Rest_32CExpr(const char *const data, const size_t len, const uint32_t hash) noexcept
+constexpr inline uint32_t Murmur3Rest_32CExpr(const char *const data, const uint32_t len, const uint32_t hash) noexcept
 {
 	return ((len & 3) == 3) ? Murmur3Tail3_32CExpr(&data[len - 3], len, hash) :
 		((len & 3) == 2) ? Murmur3Tail2_32CExpr(&data[len - 2], len, hash, 0) :
@@ -71,7 +68,7 @@ constexpr inline uint32_t Murmur3Rest_32CExpr(const char *const data, const size
 		Murmur3Last_32CExpr(len, hash);
 };
 
-constexpr inline uint32_t Murmur3Load_32CExpr(const char *const data, const size_t i) noexcept
+constexpr inline uint32_t Murmur3Load_32CExpr(const char *const data, const uint32_t i) noexcept
 {
 	return static_cast<uint32_t>(data[(i * sizeof(uint32_t)) + 3]) << 24 |
 		static_cast<uint32_t>(data[(i * sizeof(uint32_t)) + 2]) << 16 |
@@ -85,7 +82,7 @@ constexpr inline uint32_t Murmur3Update_32CExpr(const uint32_t hash, const uint3
 		0x1b873593 * Murmur3Rotate_32CExpr(0xcc9e2d51 * update, 15)), 13));
 };
 
-constexpr inline uint32_t Murmur3Loop_32CExpr(const char *const data, const size_t len, const uint32_t hash, const size_t i = 0) noexcept
+constexpr inline uint32_t Murmur3Loop_32CExpr(const char *const data, const uint32_t len, const uint32_t hash, const uint32_t i = 0) noexcept
 {
 	return (i < (len / 4)) ?
 		Murmur3Loop_32CExpr(data, len,
@@ -93,7 +90,7 @@ constexpr inline uint32_t Murmur3Loop_32CExpr(const char *const data, const size
 		Murmur3Rest_32CExpr(data, len, hash);
 };
 
-constexpr inline uint32_t Murmur3_32(const char *const key, const size_t length, const uint32_t seed) noexcept
+constexpr inline uint32_t Murmur3_32(const char *const key, const uint32_t length, const uint32_t seed) noexcept
 {
 	return Murmur3Loop_32CExpr(key, length, seed);
 };
